% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OUProcess.R
\name{densityOU}
\alias{densityOU}
\title{Calculate the density of OU process for a given (x,t) pair}
\usage{
densityOU(x, t, x0, mu, sigma, nu, log = FALSE)
}
\arguments{
\item{x}{Vector of trajectory points to calculate density of}

\item{t}{Vector of time points associated with x}

\item{x0}{Initial value of the process}

\item{mu}{Mean}

\item{sigma}{Standard deviation}

\item{nu}{Decay rate}

\item{log}{Return log(p)}
}
\description{
Calculate the density of OU process for a given (x,t) pair
}
