% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OUProcess.R
\name{expectedValOU}
\alias{expectedValOU}
\title{Expected value of OU process at times t starting at x0 at time 0}
\usage{
expectedValOU(x0, t, mu, nu)
}
\arguments{
\item{x0}{Initial value of the process}

\item{t}{Vector of time points to evaluate the trajectory at}

\item{mu}{Mean}

\item{nu}{Decay rate}

\item{sigma}{Standard deviation}
}
\description{
Expected value of OU process at times t starting at x0 at time 0
}

