% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Date_Utilities.R
\name{getDayDate}
\alias{getDayDate}
\title{Convert date from year + fraction to normal date (day, month year)}
\usage{
getDayDate(date)
}
\description{
Does not work for years with no decimal (should be December 31 of the previous year)
}
