% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logfile_Utilities.R
\name{getHPD}
\alias{getHPD}
\title{Get HPD of a posterior sample}
\usage{
getHPD(data, alpha = 0.05)
}
\arguments{
\item{data}{The samples from the posterior.}

\item{alpha}{The confidence level.}
}
\value{
c(lower, median, upper)
}
\description{
Uses Chen and Shao algorithm as implemented in boa package.
}
