% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logfile_Utilities.R
\name{getMatrixHPD}
\alias{getMatrixHPD}
\title{Get HPD of a matrix of values e.g. after gridding a skyline
Assumes by default that each row represents a posterior sample and each column an interval}
\usage{
getMatrixHPD(datamat, margin = 2, alpha = 0.05)
}
\description{
Get HPD of a matrix of values e.g. after gridding a skyline
Assumes by default that each row represents a posterior sample and each column an interval
}

