% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Date_Utilities.R
\name{getMonths}
\alias{getMonths}
\title{Return start and end date, with all months in between
From start to end, breaking at the 1st of every month in-between}
\usage{
getMonths(start, end)
}
\description{
Return start and end date, with all months in between
From start to end, breaking at the 1st of every month in-between
}

