% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPrior.R, R/SmoothingPriorPlot.R
\name{getPrior}
\alias{getPrior}
\title{Return a function with parameters that can be evaluated using eval function}
\usage{
getPrior(distr, ...)

getPrior(distr, ...)
}
\description{
Return a function with parameters that can be evaluated using eval function

Return a function with parameters that can be evaluated using eval function
}

