% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logfile_Utilities.R
\name{getSkylineSubset}
\alias{getSkylineSubset}
\title{Extract all matching parameters from the logfile
e.g. if par="R0" extract (R0s.1 R0s.2 R0s.3 etc.)}
\usage{
getSkylineSubset(logfile, par)
}
\description{
par needs to be at the start of the string
}

