% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotLikelihoodSurface.R
\name{getSurface}
\alias{getSurface}
\title{Return datastructure with the meshed likelihood surface as well as the maximum}
\usage{
getSurface(xpar, ypar, lk, crop = TRUE)
}
\description{
Does not assume par1 and par2 are ordered (so the code is not very efficient)
}

