% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tree_Utilities.R
\name{getTreeIntervals}
\alias{getTreeIntervals}
\title{Calculate the branching times and leaf heights of a tree}
\usage{
getTreeIntervals(tree, reverse = FALSE, includeRoot = TRUE)
}
\arguments{
\item{tree}{The tree to calculate the times on}

\item{reverse}{If true return time from the present, otherwise time from the root (default=FALSE)}

\item{includeRoot}{If true include the root in the time estimation (only has an effect if reverseTime is false) (default=TRUE)}
}
\description{
DOESN'T SEEM TO WORK...
Branching events labelled by a 1, leaves (sampling/death) by a 0
}
\details{
(based on getx() function by Tanja Stadler)
}

