% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Date_Utilities.R
\name{getWeeks}
\alias{getWeeks}
\title{From the first week BEFORE start up to end (strict)}
\usage{
getWeeks(start, end, inclusive = TRUE, strictend = FALSE)
}
\arguments{
\item{strictend:}{If TRUE the last value is end, if FALSE it is the first week after end.}

\item{inclusive:}{If TRUE then the weeks start and end m}
}
\description{
Weeks start on Sunday
}
