% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SkylinePlot.R
\name{gridSkyline}
\alias{gridSkyline}
\title{Interpolates skyline on a time grid (returns a matrix)
Assumes that generations are rows and skyline variables are columns (in order)
Every generation has the skyline from times[1] to origin[i] (or origin[i] to times[1] if reverse=TRUE)}
\usage{
gridSkyline(skyline, origin, times, reverse = FALSE)
}
\description{
Interpolates skyline on a time grid (returns a matrix)
Assumes that generations are rows and skyline variables are columns (in order)
Every generation has the skyline from times[1] to origin[i] (or origin[i] to times[1] if reverse=TRUE)
}

