% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logfile_Utilities.R
\name{gridSkylineDates}
\alias{gridSkylineDates}
\title{Grid the skyline between from and to}
\usage{
gridSkylineDates(skyline, origin, enddate, from, to = NA,
  intervals = "weeks", reverse = FALSE)
}
\arguments{
\item{enddate:}{End date of the skyline (most recent sample in the tree)}

\item{from:}{Start of gridded skyline}

\item{to:}{End of gridded skyline}

\item{intervals:}{Number of intervals between from and to, or 'weeks' or 'months'}
}
\description{
enddate, from and to should all be date objects or in the format "yyyy-mm-dd"
}
\details{
Assumes dates in BEAST are in years
}

