% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkylinePlot.R
\name{gridSkylineVec}
\alias{gridSkylineVec}
\title{Version using apply, but it is slower for some enigmatic reason}
\usage{
gridSkylineVec(skyline, origin, times, reverse = FALSE)
}
\description{
Version using apply, but it is slower for some enigmatic reason
}

