% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OUProcess.R
\name{likOU}
\alias{likOU}
\title{Calculates the likelihood of observing a trajectory under a given
OU-process}
\usage{
likOU(x, t, mu, sigma, nu)
}
\arguments{
\item{x}{= (x0, x1, ..., xn)}

\item{t}{= (t0, t2, ..., tn)}

\item{mu}{Mean}

\item{sigma}{Standard deviation}

\item{nu}{Decay rate}
}
\description{
Likelihood calculation follows from memorylessness of a Markov chain, 
so it's just a product of Gaussians
}
