% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkylinePlot.R
\name{paddedrange}
\alias{paddedrange}
\title{Get range and pad by some percentage of the range on either side
Good for plot limits}
\usage{
paddedrange(y, pad = 0.1)
}
\description{
Get range and pad by some percentage of the range on either side
Good for plot limits
}

