% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Figure_Utilities.R
\name{pal.dark}
\alias{pal.dark}
\title{Palette of diverging dark colours (Paired palette from RColorBrewer)}
\usage{
pal.dark(col, alpha = 1)
}
\description{
Palette of diverging dark colours (Paired palette from RColorBrewer)
}

