% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Figure_Utilities.R
\name{pal.light}
\alias{pal.light}
\title{Palette of diverging light colours (Paired palette from RColorBrewer)}
\usage{
pal.light(col, alpha = 1)
}
\description{
Palette of diverging light colours (Paired palette from RColorBrewer)
}

