% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPDPlot.R
\name{plotBeanPlot}
\alias{plotBeanPlot}
\title{Plot beans}
\usage{
plotBeanPlot(data, hpdlines = c(0.05, 0.5), linewidths = c(0.5, 0.75),
  maxwidth = 0.5, hpdonly = TRUE, plotmedian = TRUE,
  col = pal.dark(cgray, 0.25), border = pal.dark(black), side, ...)
}
\arguments{
\item{data}{List with each element being a vector that represents a posterior sample
Or a matrix, with each column being a  posterior sample}
}
\description{
Plot beans
}
