% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotLikelihoodSurface.R
\name{plotCrossSection}
\alias{plotCrossSection}
\title{Plot cross section of the likelihood surface of par1, for given values of par2
Logfile is a BEAST2 log file}
\usage{
plotCrossSection(par1, par2, lk, lk.upper = NULL, lk.lower = NULL,
  par1.truth = NA, par2.truth = NA, col = pal.dark(cblue),
  fill = pal.dark(cblue, 0.5), xlims = NULL, ylims = NULL,
  par1.label = NA, par2.label = NA, lk.label = NA, sections = -1, ...)
}
\description{
Plot cross section of the likelihood surface of par1, for given values of par2
Logfile is a BEAST2 log file
}
