% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPDPlot.R
\name{plotJellyBeanHPDs}
\alias{plotJellyBeanHPDs}
\title{Get HPDs and plot jelly beans}
\usage{
plotJellyBeanHPDs(data, maxwidth = 0.5, hpdonly = FALSE,
  plotmedian = TRUE, plotinterquartile = TRUE, col = pal.dark(cblue),
  border = NA, side = NA)
}
\arguments{
\item{data}{List with each element being a vector that represents a posterior sample
Or a matrix, with each column being a  posterior sample}
}
\description{
Get HPDs and plot jelly beans
}

