% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BeanPlot.R
\name{plotJellyBeans}
\alias{plotJellyBeans}
\title{Plot jelly beans}
\usage{
plotJellyBeans(data, maxwidth = 0.5, plotOutliers = TRUE,
  col = pal.dark(cblue), border = NA, side = NA)
}
\arguments{
\item{y}{is a matrix with 5 rows}
}
\description{
Plot jelly beans
}

