% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OUProcessPlot.R
\name{plotOUDensity}
\alias{plotOUDensity}
\title{Plot the density of an OU-process with parameters x0, mu, sigma, nu after time-interval
dt, at the range of values given by xrange.}
\usage{
plotOUDensity(xrange, dt, x0, mu, sigma, nu, col = pal.dark(cblue),
  col.alpha = 0.25, new = TRUE, plotMedian = FALSE, plotX0 = TRUE,
  plotGrid = TRUE, ...)
}
\description{
Density is only plotted between the 0.025 and 0.975 quantiles.
}
