% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OUProcessPlot.R
\name{plotOUDensityEmpirical}
\alias{plotOUDensityEmpirical}
\title{Plot the density of ntraj OU-processes simulated to time dt, using priors
x0_prior, mu_prior, sigma_prior and nu_prior on the OU-process parameters
(fixed values are also supported).}
\usage{
plotOUDensityEmpirical(xrange, dt, x0_prior, mu_prior, sigma_prior,
  nu_prior, nTraj = 1000, nSteps = 100, col = pal.dark(cblue),
  col.alpha = 0.25, bw = "sj", new = TRUE, plotMedian = FALSE,
  plotX0 = TRUE, plotGrid = TRUE, ...)
}
\description{
Density is only plotted in the 95% HPD interval, at the points in xrange.
}
