% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OUProcessPlot.R
\name{plotOUProcessFlexibility}
\alias{plotOUProcessFlexibility}
\title{Plot a summary of the flexibility of an OU-process with parameters mu, sigma and nu, starting 
from different x0 values, over a time period of dt}
\usage{
plotOUProcessFlexibility(xlim, dt, x0, mu, sigma, nu, nSteps = 100)
}
\description{
Answers the questions: 
- How big is the 95% CI after one time interval?
}
