% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OUProcessPlot.R
\name{plotOUProcessFlexibilityEmpirical}
\alias{plotOUProcessFlexibilityEmpirical}
\title{Plot a summary of the flexibility of an OU-process with parameters sampled from priors 
mu_prior, sigma_prior and nu_prior, starting from different x0 values, over a time period 
of dt}
\usage{
plotOUProcessFlexibilityEmpirical(xlim, dt, x0, mu_p, sigma_p, nu_p,
  nSteps = 100)
}
\description{
Answers the questions: 
- How big is the 95% CI after one time interval?
}
