% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OUProcessPlot.R
\name{plotOUProcessHPDEmpirical}
\alias{plotOUProcessHPDEmpirical}
\title{Plot empirical 95% HPD interval and median of OU-process, 
from simulating nTraj replicates, which can also be plotted}
\usage{
plotOUProcessHPDEmpirical(x0_prior, t, mu_prior, sigma_prior, nu_prior,
  nTraj = 1000, plotTraj = TRUE, ylim = c(0, 5), ...)
}
\description{
Instead of specifying constant parameter values, prior functions can also
be specified, in which case they will be sampled for simulating trajectories
}
