% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmoothingPriorPlot.R
\name{plotOUProcessPriors}
\alias{plotOUProcessPriors}
\title{Parameters may be a prior or a constant
Use getPrior to quote the prior function to be passed with parameters}
\usage{
plotOUProcessPriors(x0_prior, t, mu_prior, sigma_prior, nu_prior, ...)
}
\description{
Should probably rewrite this using do.call()
}
\details{
use par("usr") to get clipping coordinates and then use clip() to set the clipping region
}

