% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPrior.R
\name{plotQuantileGradient}
\alias{plotQuantileGradient}
\title{Plot the quantiles for OU or BM process as a gradient across the timepoints}
\usage{
plotQuantileGradient(t, quantiles, colidx)
}
\arguments{
\item{t}{Vector of time points to plot quantiles at}

\item{quantiles}{The quantiles as determined by quantilesOU or quantilesBM}

\item{col}{A colour constant, e.g. cblue or cgreen (integer from 1-10)}
}
\description{
Plot the quantiles for OU or BM process as a gradient across the timepoints
}
