% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPDPlot.R
\name{plotRoundedBars}
\alias{plotRoundedBars}
\title{Plot a rounded bar (jelly bean)}
\usage{
plotRoundedBars(x, y1, y2, width, nv = 100, plotAlways = FALSE,
  flip = FALSE, side = NA, ...)
}
\arguments{
\item{x}{x-coordinate of the bar}

\item{y1}{Upper coordinate of the bar}

\item{y2}{Lower coordinate of the bar}

\item{width}{Width of the bar}

\item{nv}{Number of vertices to use for drawing the circle}

\item{plotAlways}{Plot even if the proportions will make it look bad}

\item{flip}{Plot horizontally (x is interpreted as y, and y1, y2 as x1, x2)}

\item{side}{Plot only half of the bar. (If side = 2 plot the left half, if side = 4 plot the right half)}
}
\description{
Plot a rounded bar (jelly bean)
}

