% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SkylinePlot.R
\name{plotSkyline}
\alias{plotSkyline}
\title{new: create a new set of axes
add: add to the current plot (do not create a new plotting device)}
\usage{
plotSkyline(times, skyline_mat, type = "smooth", col = pal.dark(cblack),
  fill = pal.dark(cgray, 0.25), traces = 1000, new = TRUE, add = FALSE,
  ...)
}
\arguments{
\item{type}{Type of skyline to plot.

            "smooth" : If skyline_mat contains the HPDs plot a smooth line and polygon (type='l')

            "step"   : If skyline_mat contains the HPDs plot a stepped line and polygon (type='S')

            "traces" : Plot every row in the matrix as a stepped line (type='S')}

\item{traces}{Number of traces to draw if type="traces"}

\item{new}{Create a new set of axes (the skyline is fitted to the plotting devie)}

\item{add}{Add to the current plot (do not create a new plotting device)}

\item{...}{Parameters passed to plotting function}
}
\description{
assume each column is a different time point (ncol(skylinemat) == length(times) must be fulfilled)
}

