% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkylinePlot.R
\name{plotSkylinePretty}
\alias{plotSkylinePretty}
\title{Plot a pretty skyline. Contains default options for drawing axes to the plot}
\usage{
plotSkylinePretty(times, skyline_mat, type = "smooth", traces = 1000,
  col = pal.dark(cblack), fill = pal.dark(cgray, 0.25),
  col.axis = NA, lwd = 1, lty = 1, xaxis = TRUE, yaxis = TRUE,
  xlab = "", ylab = "", xline = 1, yline = 1, xticks = NULL,
  yticks = NULL, xticklabels = NULL, xlims = NULL, ylims = NULL,
  xgrid = FALSE, ygrid = FALSE, gridcol = pal.light(cgray),
  cex.label = 1.4, cex.axis = 1, axispadding = 0, side = 2, ...)
}
\arguments{
\item{xline}{Line to draw x-axis label on}

\item{yline}{Line to draw y-axis label on}

\item{side}{Side to draw the y-axis}
}
\description{
When plotting the y-axis on the right need to ensure that the plot margins are big enough!
See plotSkyline help for documentation on other parameters!
}
