% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMProcess.R
\name{quantilesBM}
\alias{quantilesBM}
\title{Calculate quantiles of BM process at different timepoints}
\usage{
quantilesBM(p, x0, t)
}
\arguments{
\item{x0}{Initial value of the process}

\item{t}{Vector of time points to evaluate the trajectory at}
}
\description{
Calculate quantiles of BM process at different timepoints
}
