% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OUProcess.R
\name{quantilesOU}
\alias{quantilesOU}
\title{Calculate quantiles of OU process at different timepoints}
\usage{
quantilesOU(p, x0, t, mu, sigma, nu)
}
\arguments{
\item{x0}{Initial value of the process}

\item{t}{Vector of time points to evaluate the trajectory at}

\item{mu}{Mean}

\item{sigma}{Standard deviation}

\item{nu}{Decay rate}
}
\description{
Calculate quantiles of OU process at different timepoints
}

