% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logfile_Utilities.R
\name{readLogfile}
\alias{readLogfile}
\title{Read in BEAST2 logfile
maxsamples is only there for testing (load quicker)}
\usage{
readLogfile(filename, burnin = 0.1, maxsamples = -1)
}
\description{
Read in BEAST2 logfile
maxsamples is only there for testing (load quicker)
}

