% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMProcess.R
\name{simulateBM}
\alias{simulateBM}
\title{Simulate BM trajectory}
\usage{
simulateBM(x0, t, k = 1)
}
\arguments{
\item{x0}{Initial value of the process}

\item{t}{Vector of time points to evaluate the trajectory at}
}
\description{
Simulate BM trajectory
}
