% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OUProcess.R
\name{simulateDiscreteOU}
\alias{simulateDiscreteOU}
\title{Functions for simulating Ornstein-Uhlenbeck processes and calculating densities and likelihoods}
\usage{
simulateDiscreteOU(x0, t, mu, sigma, nu)
}
\arguments{
\item{x0}{Initial value of the process}

\item{t}{Vector of time points to evaluate the trajectory at}

\item{mu}{Mean}

\item{sigma}{Standard deviation}

\item{nu}{Decay rate}
}
\description{
Simulate OU trajectory using the Euler-Maruyama method (inaccurate)
}
\details{
Function assumes equidistant time intervals
}
