% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OUProcess.R
\name{simulateOUslow}
\alias{simulateOUslow}
\title{Simulate OU trajectory using the analytical solution obtained
using a scaled time-transformed Wiener process}
\usage{
simulateOUslow(x0, t, mu, sigma, nu)
}
\arguments{
\item{x0}{Initial value of the process}

\item{t}{Vector of time points to evaluate the trajectory at}

\item{mu}{Mean}

\item{sigma}{Standard deviation}

\item{nu}{Decay rate}
}
\description{
Slow unvectorized version
}

