% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OUProcess.R
\name{standardDevOU}
\alias{standardDevOU}
\title{Standard deviation of OU process at times t starting at x0 at time 0}
\usage{
standardDevOU(x0, t, mu, sigma, nu)
}
\arguments{
\item{x0}{Initial value of the process}

\item{t}{Vector of time points to evaluate the trajectory at}

\item{mu}{Mean}

\item{sigma}{Standard deviation}

\item{nu}{Decay rate}
}
\description{
Standard deviation of OU process at times t starting at x0 at time 0
}
