% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Logfile_Utilities.R
\name{getMatrixHPD}
\alias{getMatrixHPD}
\title{Get HPD of a matrix of values e.g. a skyline
Uses Chen and Shao algorithm as implemented in boa package.}
\usage{
getMatrixHPD(datamat, margin = 2, alpha = 0.05)
}
\arguments{
\item{alpha}{The confidence level.}

\item{data}{The samples from the posterior. Assumes by default that each row represents a posterior sample and each column a parameter (interval).}
}
\value{
3xn matrix where each column is c(lower, median, upper)
}
\description{
Get HPD of a matrix of values e.g. a skyline
Uses Chen and Shao algorithm as implemented in boa package.
}

