% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPrior.R
\name{plotBMProcessPriors}
\alias{plotBMProcessPriors}
\title{Parameters may be a prior or a constant}
\usage{
plotBMProcessPriors(x0_prior, t, ...)
}
\description{
Use getPrior to quote the prior function to be passed with parameters
}
