% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPrior.R, R/SmoothingPriorPlot.R
\name{plotBMProcessPriors}
\alias{plotBMProcessPriors}
\title{Parameters may be a prior or a constant
Use getPrior to quote the prior function to be passed with parameters}
\usage{
plotBMProcessPriors(x0_prior, t, ...)

plotBMProcessPriors(x0_prior, t, ...)
}
\description{
Parameters may be a prior or a constant
Use getPrior to quote the prior function to be passed with parameters

Parameters may be a prior or a constant
Use getPrior to quote the prior function to be passed with parameters
}

