% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPDPlot.R
\name{plotJellyBeans}
\alias{plotJellyBeans}
\title{Plot jelly beans}
\usage{
plotJellyBeans(x, y, maxwidth = 0.5, plotmedian = TRUE,
  plotinterquartile = TRUE, col = pal.dark(cblue), border = NA,
  side = NA)
}
\arguments{
\item{x}{locations to plot jellybeans}

\item{y}{matrix with 5 rows and n columns}
}
\description{
Plot jelly beans
}
