% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPrior.R
\name{plotOUProcessPriors}
\alias{plotOUProcessPriors}
\title{Parameters may be a prior or a constant
Use getPrior to quote the prior function to be passed with parameters}
\usage{
plotOUProcessPriors(x0_prior, t, mu_prior, sigma_prior, nu_prior, ...)
}
\description{
Should probably rewrite this using do.call()
}
