% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkylinePlot.R
\name{plotSkyline}
\alias{plotSkyline}
\title{new: create a new set of axes
add: add to the current plot (do not create a new plotting device)}
\usage{
plotSkyline(skyline_mat, times, type = "", col = pal.dark(cblack),
  fill = pal.dark(cgray, 0.25), new = TRUE, add = FALSE, ...)
}
\description{
assume each column is a different time point (ncol(skylinemat) == length(times) must be fulfilled)
}

