% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkylinePlot.R
\name{plotSkyline}
\alias{plotSkyline}
\title{Plot a Skyline.}
\usage{
plotSkyline(times, skyline_mat, type = "smooth", traces = 1000,
  col = pal.dark(cblack), fill = pal.dark(cgray, 0.25), lwd = 1,
  lty = 1, new = TRUE, add = FALSE, xlims = NULL, ylims = NULL,
  ...)
}
\arguments{
\item{times}{The time points to draw the skyline at. length(times) should be equal to ncol(skyline_mat) + 1.
(Because there are length(times) intervals and the last interval needs an end-time).
If length(times) == ncol(skyline_mat) the last interval is arbitrarily given the same length as 
the second last interval.}

\item{skyline_mat}{}

\item{type}{Type of skyline to plot.

            "smooth" : If skyline_mat contains the HPDs plot a smooth line and polygon (type='l')
            
            "step"   : If skyline_mat contains the HPDs plot a stepped line and polygon (type='S')
            
            "lines" : Plot every row in the matrix as a smooth line (type='l')
            
            "steplines" : Plot every row in the matrix as a stepped line (type='S')}

\item{traces}{Number of traces to draw if type="traces"}

\item{new}{Create a new set of axes (the skyline is fitted to the plotting devie)}

\item{add}{Add to the current plot (do not create a new plotting device)}

\item{...}{Parameters passed to plotting function}
}
\description{
Assume each column is a different time point (ncol(skylinemat) == length(times) must be fulfilled)
}
