% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SkylinePlot.R
\name{plotSkyline}
\alias{plotSkyline}
\title{Plot a Skyline.}
\usage{
plotSkyline(times, skyline_mat, type = "smooth", traces = 1000,
  col = pal.dark(cblack), fill = pal.dark(cgray, 0.25), new = TRUE,
  add = FALSE, xlims = NULL, ylims = NULL, ...)
}
\arguments{
\item{type}{Type of skyline to plot.

            "smooth" : If skyline_mat contains the HPDs plot a smooth line and polygon (type='l')

            "step"   : If skyline_mat contains the HPDs plot a stepped line and polygon (type='S')

            "lines" : Plot every row in the matrix as a smooth line (type='l')

            "steplines" : Plot every row in the matrix as a stepped line (type='S')}

\item{traces}{Number of traces to draw if type="traces"}

\item{new}{Create a new set of axes (the skyline is fitted to the plotting devie)}

\item{add}{Add to the current plot (do not create a new plotting device)}

\item{...}{Parameters passed to plotting function}
}
\description{
Assume each column is a different time point (ncol(skylinemat) == length(times) must be fulfilled)
}

