% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SkylinePlot.R
\name{plotSkylinePretty}
\alias{plotSkylinePretty}
\title{Plot a pretty skyline}
\usage{
plotSkylinePretty(times, skyline_mat, type = "smooth", traces = 1000,
  col = cgray, col.axes = FALSE, side = 2, xlab = "", ylab = "",
  xline = 1, yline = 1, xticks = NULL, yticks = NULL, xlims = NULL,
  ylims = NULL, ...)
}
\arguments{
\item{side}{Side to draw the y-axis}

\item{xline}{Line to draw x-axis label on}

\item{yline}{Line to draw y-axis label on}
}
\description{
Plot a pretty skyline
}

