% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logfile_Utilities.R
\name{revMatrix}
\alias{revMatrix}
\title{Reverses each column (margin=2) or row (margin=1) of a matrix}
\usage{
revMatrix(data, margin = 1)
}
\description{
Reverses each column (margin=2) or row (margin=1) of a matrix
}

